frontend/api/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
boto3
cryptography
mangum
